"Delete Quotes Functionality. #9" issue resolved

+++ src/App.jsx
@@ -6,10 +6,14 @@ import axios from 'axios';
 import { ToastContainer } from 'react-toastify';
 import Search from './components/Search';
 import Header from './components/Header';
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 //API Details
 // 20220912223415
 // https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app
+
+const ID = nanoid()
 
 function App() {
 
@@ -18,7 +22,7 @@ function App() {
 
   const [notes, setNotes] = useState([
     {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: 'this is the note pad app text',
@@ -47,7 +51,7 @@ function App() {
   const addNote = (text) => {
     const date = new Date();
     const newNote = {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: text,
@@ -67,37 +71,48 @@ function App() {
   const deleteNote = (id) => {
     const newNotes = notes.filter((note) => note.id !== id);
     //delete operation
-    axios.delete(
-      `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
-      setNotes(newNotes)
-    );
+    if (id == ID)
+      axios.delete(
+        `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
+        setNotes(newNotes)
+      )
+    else
+      toast('📋 This is not YourQutoes', {
+        position: 'top-center',
+        autoClose: 2000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+      })
   };
 
   return (
     <>
-     <div className={`${darkMode && 'dark-mode'}`}>
-     <Header handleToggleDarkMode={setDarkMode} />
-    <div className="container">
-      <Search handleSearchNote={setSearchText} />
-      <CardsList
-        notes={notes.filter((note)=> note.text.toUpperCase().includes(searchText.toLocaleUpperCase()))}
-        handleAddNote={addNote}
-        handleDeleteNote={deleteNote}
-      />
-      <ToastContainer
-        position="top-center"
-        autoClose={2000}
-        hideProgressBar={false}
-        newestOnTop={false}
-        closeOnClick
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover
-      />
-      <ToastContainer />
-    </div>
-     </div>
+      <div className={`${darkMode && 'dark-mode'}`}>
+        <Header handleToggleDarkMode={setDarkMode} />
+        <div className="container">
+          <Search handleSearchNote={setSearchText} />
+          <CardsList
+            notes={notes.filter((note) => note.text.toUpperCase().includes(searchText.toLocaleUpperCase()))}
+            handleAddNote={addNote}
+            handleDeleteNote={deleteNote}
+          />
+          <ToastContainer
+            position="top-center"
+            autoClose={2000}
+            hideProgressBar={false}
+            newestOnTop={false}
+            closeOnClick
+            rtl={false}
+            pauseOnFocusLoss
+            draggable
+            pauseOnHover
+          />
+          <ToastContainer />
+        </div>
+      </div>
     </>
   );
 }
